+++ proto/src/lib.rs
@@ -8,6 +8,9 @@
 #![forbid(unsafe_code)]
 #![doc(html_root_url = "https://docs.rs/ibc-proto/0.4.0")]
 
+/// The version (commit hash) of the Cosmos SDK used when generating this library.
+pub const COSMOS_SDK_VERSION: &str = include_str!("prost/COSMOS_SDK_COMMIT");
+
 pub mod cosmos {
     pub mod base {
         pub mod abci {