+++ CHANGELOG.md
@@ -13,6 +13,15 @@ Special thanks to external contributors for this release: @CharlyCst ([#347]).
 - [relayer-cli]
     - Merge light clients config in relayer config and add commands to add/remove light clients ([#348])
     - CLI for client update message ([#277])
+- [proto-compiler] Refactor and allow specifying a commit at which the Cosmos SDK should be checked out ([#366])
+- [ibc-proto] Refactor and allow specifying a commit at which the Cosmos SDK should be checked out ([#366])
+
+[#274]: https://github.com/informalsystems/ibc-rs/issues/274
+[#332]: https://github.com/informalsystems/ibc-rs/issues/332
+[#335]: https://github.com/informalsystems/ibc-rs/pulls/335
+[#336]: https://github.com/informalsystems/ibc-rs/issues/336
+[#348]: https://github.com/informalsystems/ibc-rs/pulls/348
+[#366]: https://github.com/informalsystems/ibc-rs/issues/366
 
 ### IMPROVEMENTS